more type cleanup

+++ src/cpu.h
@@ -43,7 +43,7 @@ struct aldo_mos6502 {
 
     // Datapath: abstract representation of instruction fetching,
     // execution, and signaling.
-    int8_t t;           // Instruction sequence cycle (T0, T1, T2...)
+    int t;              // Instruction sequence cycle (T0, T1, T2...)
     enum aldo_sigstate
         irq,            // IRQ detection latch
         nmi,            // NMI detection latch
